Rollback
This is a premature commit but I reached a point in "beautifying" code that I destroyed functionality and couldn't figure out why. So I have my "NEW" file with these changes, but the original source modified just enough to get it working. (there are still bugs related to variable placement/scope)

On a side note, I updated README to mention where to get team files.